requirements.txt
streamlit==1.51.0
firebase_admin==7.1.0
gspread==6.2.1
oauth2client==4.1.3
boto3==1.41.1
PyYAML==6.0.3
platformdirs==4.5.0